Score logic, drawBoard out of class, export class
